move method to common util
