+++ pkg/kyverno/common/common.go
@@ -1,11 +1,8 @@
 package common
 
 import (
-	"bufio"
-	"bytes"
 	"encoding/json"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"os"
 	"path/filepath"
@@ -18,7 +15,7 @@ import (
 	"github.com/nirmata/kyverno/pkg/kyverno/sanitizedError"
 	"github.com/nirmata/kyverno/pkg/openapi"
 	"github.com/nirmata/kyverno/pkg/policymutation"
-	"k8s.io/apimachinery/pkg/util/yaml"
+	"github.com/nirmata/kyverno/pkg/utils"
 	log "sigs.k8s.io/controller-runtime/pkg/log"
 )
 
@@ -53,7 +50,7 @@ func GetPolicies(paths []string) (policies []*v1.ClusterPolicy, error error) {
 
 			policies = append(policies, policiesFromDir...)
 		} else {
-			getPolicies, getErrors := GetPolicy(path)
+			getPolicies, getErrors := utils.GetPolicy(path)
 			var errString string
 			for _, err := range getErrors {
 				if err != nil {
@@ -71,62 +68,6 @@ func GetPolicies(paths []string) (policies []*v1.ClusterPolicy, error error) {
 	}
 
 	return policies, nil
-}
-
-// GetPolicy - Extracts policies from a YAML
-func GetPolicy(path string) (clusterPolicies []*v1.ClusterPolicy, errors []error) {
-	file, err := ioutil.ReadFile(path)
-	if err != nil {
-		errors = append(errors, fmt.Errorf(fmt.Sprintf("failed to load file: %v. error: %v", path, err)))
-		return clusterPolicies, errors
-	}
-
-	policies, err := SplitYAMLDocuments(file)
-	if err != nil {
-		errors = append(errors, err)
-		return clusterPolicies, errors
-	}
-
-	for _, thisPolicyBytes := range policies {
-		policyBytes, err := yaml.ToJSON(thisPolicyBytes)
-		if err != nil {
-			errors = append(errors, fmt.Errorf(fmt.Sprintf("failed to convert json. error: %v", err)))
-			continue
-		}
-
-		policy := &v1.ClusterPolicy{}
-		if err := json.Unmarshal(policyBytes, policy); err != nil {
-			errors = append(errors, fmt.Errorf(fmt.Sprintf("failed to decode policy in %s. error: %v", path, err)))
-			continue
-		}
-
-		if policy.TypeMeta.Kind != "ClusterPolicy" {
-			errors = append(errors, fmt.Errorf(fmt.Sprintf("resource %v is not a cluster policy", policy.Name)))
-			continue
-		}
-		clusterPolicies = append(clusterPolicies, policy)
-	}
-
-	return clusterPolicies, errors
-}
-
-// SplitYAMLDocuments reads the YAML bytes per-document, unmarshals the TypeMeta information from each document
-// and returns a map between the GroupVersionKind of the document and the document bytes
-func SplitYAMLDocuments(yamlBytes []byte) (policies [][]byte, error error) {
-	buf := bytes.NewBuffer(yamlBytes)
-	reader := yaml.NewYAMLReader(bufio.NewReader(buf))
-	for {
-		// Read one YAML document at a time, until io.EOF is returned
-		b, err := reader.Read()
-		if err == io.EOF || len(b) == 0 {
-			break
-		} else if err != nil {
-			return policies, fmt.Errorf("unable to read yaml")
-		}
-
-		policies = append(policies, b)
-	}
-	return policies, error
 }
 
 //GetPoliciesValidation - validating policies
